+++ app/views/TodoList.scala.html
@@ -1,6 +1,6 @@
-@(todos: Seq[DisplayTodo])(implicit request: RequestHeader, messagesProvider: MessagesProvider)
-
+@import models.todos._
 @import helper._
+@(todos: Seq[DisplayTodo])(implicit request: RequestHeader, messagesProvider: MessagesProvider)
 
 <b>todo一覧</b>
 @if(!todos.isEmpty){
@@ -21,7 +21,7 @@
     <td><pre>@todo.body</pre></td>
       <td>@todo.state.name</td>
       <td style="color: @(todo.category.v.color.code)">@todo.category.v.name</td>
-      <td><button name="id", value="@todo.id">変更</button></td>
+      <td><button name="id" value="@todo.id">変更</button></td>
       <td>
     }
     @form(routes.HomeController.delete(todo.id)){
